move SworkerInterface and MarketInterface to primitives
